+++ AmazonChimeSDK/AmazonChimeSDK/ingestion/IngestionPayload.swift
@@ -22,8 +22,6 @@ import Foundation
     public let retryCount: Int?
     public let videoInputErrorMessage: String?
     public let audioInputErrorMessage: String?
-    public let audioAccessErrorMessage: String?
-    public let videoAccessErrorMessage: String?
     public let ttl: Int64?
 
     public init(name: String,
@@ -39,8 +37,6 @@ import Foundation
                 retryCount: Int? = nil,
                 videoInputErrorMessage: String? = nil,
                 audioInputErrorMessage: String? = nil,
-                audioAccessErrorMessage: String? = nil,
-                videoAccessErrorMessage: String? = nil,
                 ttl: Int64? = nil) {
         self.name = name
         self.ts = ts
@@ -55,8 +51,6 @@ import Foundation
         self.retryCount = retryCount
         self.videoInputErrorMessage = videoInputErrorMessage
         self.audioInputErrorMessage = audioInputErrorMessage
-        self.audioAccessErrorMessage = audioAccessErrorMessage
-        self.videoAccessErrorMessage = videoAccessErrorMessage
         self.ttl = ttl
     }
 }